// sourcecode/BTrak.Packages/snova-document-management/document-module-components/document-components/src/lib/document-management/models/search-folder-model.ts
import { SearchCriteriaInputModelBase } from "./searchCriteriaInputModelBase";

export class SearchFolderModel extends SearchCriteriaInputModelBase{
    folderId: string;
    parentFolderId: string;
    storeId: string;
    folderReferenceId: string;
    folderReferenceTypeId: string;
    isArchived: boolean;
    searchText: string;
    isFromSprints: boolean;
    isTreeView: boolean;
    userStoryId: string;
}